+++ backend/tests/app/routers/test_qa.py
@@ -56,7 +56,9 @@ def test_chat_streamed(app_client, httpx_mock, patch_required_env, db_connection
     qa.stream_agent_response.return_value = streamed_response()
     mock_keycloak_user_identification(httpx_mock)
     test_settings = Settings(
-        db={"prefix": db_connection}, keycloak={"issuer": "https://great_issuer.com"}
+        db={"prefix": db_connection},
+        keycloak={"issuer": "https://great_issuer.com"},
+        accounting={"disabled": True},
     )
     app.dependency_overrides[get_settings] = lambda: test_settings
     agent_routine = Mock()