+++ backend/tests/conftest.py
@@ -35,6 +35,8 @@ def client_fixture():
         openai={
             "token": "fake_token",
         },
+        rate_limiter={"disabled": True},
+        accounting={"disabled": True},
     )
     app.dependency_overrides[get_settings] = lambda: test_settings
     yield app_client
@@ -274,4 +276,6 @@ def settings():
         openai={
             "token": "fake_token",
         },
+        rate_limiter={"disabled": True},
+        accounting={"disabled": True},
     )